add support to create new docker images from user lambda code or custom gitlab repository

+++ src/main/kotlin/cz/chobot/image_builder/bo/NetworkType.kt
@@ -10,7 +10,7 @@ data class NetworkType(
         @Id
         @GeneratedValue(strategy = GenerationType.IDENTITY)
         @Column(name = "network_type_id", unique = true, nullable = false)
-        val network_type_id: Int,
+        val network_type_id: Long,
 
         @NotNull
         @Column(name = "name", nullable = false)
